[Redis] feat: move redis 파일 추가 이동
