+++ src/main/java/com/mzc/redis/controller/UserController.java
@@ -1,6 +1,7 @@
-package com.mzc.controller;
+package com.mzc.redis.controller;
 
-import com.mzc.Service.UserService;
+import com.mzc.redis.Service.UserService;
+import com.mzc.redis.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
